Refactor tests to simulate low event counts
With this commit we refactor the profiling-related test cases so we can
simulate situations after cluster startup when not all profiling-related
indices have been created. As testing also cancellation would require
additional complexity and decreases test maintainability (need to adjust
`slack` based on the scenario which is a very low-level change), we also
separate tests for the regular case and cancellation.
